Match types between refs and coeffs
Change type of av1_warped_filter from 16-bit to 8-bit to
match how data is used naturally in the code. All values
are 8-bit anyway. This may also reduce memory usage.

+++ av1/common/warped_motion.h
@@ -33,7 +33,7 @@
 #define WARP_ERROR_BLOCK_LOG 5
 #define WARP_ERROR_BLOCK (1 << WARP_ERROR_BLOCK_LOG)
 
-extern const int16_t av1_warped_filter[WARPEDPIXEL_PREC_SHIFTS * 3 + 1][8];
+extern const int8_t av1_warped_filter[WARPEDPIXEL_PREC_SHIFTS * 3 + 1][8];
 
 DECLARE_ALIGNED(8, extern const int8_t,
                 av1_filter_8bit[WARPEDPIXEL_PREC_SHIFTS * 3 + 1][8]);
